+++ app/helpers/wilday_ui/application_helper.rb
@@ -1,5 +1,6 @@
 module WildayUi
   module ApplicationHelper
     include WildayUi::Components::ButtonHelper
+    include WildayUi::JavascriptHelper
   end
 end