+++ app/helpers/wilday_ui/javascript_helper.rb
@@ -1,0 +1,13 @@
+module WildayUi
+  module JavascriptHelper
+    def w_js(options = {})
+      javascript_include_tag(
+        "wilday_ui/index",
+        {
+          "data-turbo-track": "reload",
+          type: "module"
+        }.merge(options)
+      )
+    end
+  end
+end