+++ app/javascript/wilday_ui/components/button.js
@@ -1,5 +1,4 @@
 document.addEventListener("DOMContentLoaded", () => {
-  console.log("Button component loaded heah");
   document.querySelectorAll(".w-button").forEach((button) => {
     button.addEventListener("click", (event) => {
       if (button.disabled) {